split index file from pack
