+++ storage/src/containers/mod.rs
@@ -1,2 +1,3 @@
 pub mod append;
 pub mod packfile;
+pub mod indexfile;
